fix(web-barbies): relocate the about page api slice
