+++ Web/starter_project_barbies/components/about/ProblemsSection.tsx
@@ -1,7 +1,7 @@
 import {
   useGetProblemItemsQuery,
   useGetSolutionItemsQuery,
-} from '@/store/features/about/about-api'
+} from '@/store/about/about-api'
 
 import { ProblemItem } from '@/types/about'
 import Image from 'next/image'
